feat: 이미지 슬라이드 컴포넌트 생성 및 애니메이션 추가
