+++ src/pages/MainPage.css
@@ -1,10 +1,12 @@
-.slider-container {
+/* 이미지 슬라이드 css */
+
+.slider-img {
   text-align: center;
 }
 
 .image-wrapper {
   position: relative;
-  width: 800px;
+  width: 1280px;
   height: 300px;
   margin: 0 auto;
   overflow: hidden;
@@ -14,13 +16,24 @@
   width: 100%;
   height: 100%;
   object-fit: cover;
-  transition: opacity 0.8s ease-in-out, transform 0.8s ease-in-out; /* 애니메이션 효과 */
+  transition: opacity 0.4s ease-in-out, transform 0.4s ease-in-out; /* 애니메이션 효과 시간 변경 */
   opacity: 1;
 }
 
 .image-leaving {
   opacity: 0;
+}
+
+.image-entering {
+  opacity: 0;
+  transform: translateX(0);
+  transition: opacity 0.2s ease-in-out;
+}
+
+.image-entering-hidden {
+  opacity: 0;
   transform: translateX(-100%);
+  transition: opacity 0.4s ease-in-out;
 }
 
 .button {
@@ -43,3 +56,5 @@
 .button-next {
   right: 10px;
 }
+
+/* 이미지 슬라이드 끝 */
